#lockdown nick.penwarden
#rb none

#ROBOMERGE-OWNER: robert.manuszewski
#ROBOMERGE-AUTHOR: jeanmichel.dignard
#ROBOMERGE-SOURCE: CL 6890764 in //UE4/Main/...
#ROBOMERGE-BOT: CORE (Main -> Dev-Core) (v366-6836689)

[CL 6906312 by stefan boberg in Dev-Core branch]

+++ Engine/Source/Runtime/MovieScene/Public/Channels/MovieSceneChannelEditorData.h
@@ -45,7 +45,7 @@ struct FMovieSceneChannelMetaData
 	/** True if this channel can be collapsed onto the top level track node */
 	uint8 bCanCollapseToTrack : 1;
 	/** A sort order for this channel. Channels are sorted by this order, then by name. Groups are sorted by the channel with the lowest sort order. */
-	uint8 SortOrder;
+	uint32 SortOrder;
 	/** This channel's unique name */
 	FName Name;
 	/** Text to display on this channel's key area node */
